+++ packages/atlas/src/components/AppKV/AppKV.styles.ts
@@ -1,7 +1,7 @@
 import styled from '@emotion/styled'
 
 import { Text } from '@/components/Text'
-import { media, sizes } from '@/styles'
+import { sizes, zIndex } from '@/styles'
 
 export const Wrapper = styled.div`
   position: relative;
@@ -24,15 +24,13 @@ export const LogoWrapper = styled.div`
   flex-direction: column;
   justify-content: center;
   align-items: center;
+  z-index: ${zIndex.overlay};
 `
 
 export const Fade = styled.div`
   position: absolute;
-  inset: 0;
-  top: 40%;
+  bottom: 0;
   background: linear-gradient(180deg, rgb(24 28 32 / 0) 0%, #181c20 100%);
-
-  ${media.sm} {
-    top: 52%;
-  }
+  width: 100%;
+  height: 128px;
 `