+++ CHANGELOG.md
@@ -9,7 +9,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- Welcome dialog when content focus is present
+
 ### Changed
+
+- Styles on SignIn dialog and Studio welcome view
 
 ### Fixed